feat: adds keyboard controls & closes menu on pick
Changed to using a button element instead of a span for the trigger
element since that type of element seemed more appropriate semantically.

The closeMenu helper function is a bit hacky since I wasn't able to find
a way to add an event listener to a specific element in a v-for loop. A
preliminary search made it seem like it wasn't possible so I came up with
this for a work around. It is mostly needed to aid with keyboard controls
since when you tab off of the last element you should probably close the
menu altogether.

Finally, as for Vue 3 best practices, I'm not well versed in that yet,
but the main thing I noticed was setOption was trying to set the value
props directly which is not something you want to do. If we want to
maintain this state internally then we could create a reactive ref to
hold the currently selected hero.