refactor: turns bonus into computed property
This change allows us to simply increment each stat when the bonus
button is clicked for the selected hero. We then compute the bonus based
on the starting base total of all stats and the current total of all
stats.

This gives us the advantage of each hero having different bonuses
applied at any given time and having those changes persist when
selecting different heroes.